Fixes #123. Support out-of-tree maven builds.
Don't try to shorten paths to xsd sources. XCJ is called directly, not
via the command-line, so doing so doesn't really help. Not trying to
shorten these paths also avoids stripping a leading '/', which would
otherwise make XCJ treat the path as relative and resolve to the
current directory, which breaks out-of-tree builds.

Adds an integration test that creates a maven project in the system
tmp directory, runs a maven build there, and verifies that that maven
build was successful and did generate something.